+++ ncollide_pipeline/narrow_phase/narrow_phase.rs
@@ -3,7 +3,8 @@ use utils::data::hash_map::Entry;
 use utils::data::pair::Pair;
 use utils::data::uid_remap::{UidRemap, FastKey};
 use geometry::query::Contact;
-use narrow_phase::{CollisionAlgorithm, ContactSignal, CollisionDetector, ProximitySignal};
+use narrow_phase::{CollisionAlgorithm, ContactSignal, CollisionDetector,
+                   ProximityAlgorithm, ProximitySignal, ProximityDetector};
 use world::CollisionObject;
 use math::Point;
 
@@ -20,22 +21,24 @@ pub trait NarrowPhase<P: Point, M, T> {
               timestamp:        usize);
 
     /// Called when the broad phase detects that two objects are, or stop to be, in close proximity.
-    ///
-    /// It is expected not to be called if `self.is_proximity_allowed(...)` return `false` for the
-    /// given pair of objects.
-    fn handle_proximity(&mut self,
-                        contact_signal:   &mut ContactSignal<T>,
-                        proximity_signal: &mut ProximitySignal<T>,
-                        objects:          &UidRemap<CollisionObject<P, M, T>>,
-                        fk1:              &FastKey,
-                        fk2:              &FastKey,
-                        started:          bool);
+    fn handle_interaction(&mut self,
+                          contact_signal:   &mut ContactSignal<T>,
+                          proximity_signal: &mut ProximitySignal<T>,
+                          objects:          &UidRemap<CollisionObject<P, M, T>>,
+                          fk1:              &FastKey,
+                          fk2:              &FastKey,
+                          started:          bool);
 
     // FIXME: the fact that the return type is imposed is not as generic as it could be.
     /// Returns all the potential contact pairs found during the broad phase, and validated by the
     /// narrow phase.
     fn contact_pairs<'a>(&'a self, objects: &'a UidRemap<CollisionObject<P, M, T>>)
                          -> ContactPairs<'a, P, M, T>;
+
+    /// Returns all the potential proximity pairs found during the broad phase, and validated by
+    /// the narrow phase.
+    fn proximity_pairs<'a>(&'a self, objects: &'a UidRemap<CollisionObject<P, M, T>>)
+                         -> ProximityPairs<'a, P, M, T>;
 }
 
 /// Iterator through contact pairs.
@@ -134,3 +137,41 @@ impl<'a, P: Point, M, T> Iterator for Contacts<'a, P, M, T> {
     }
 }
 
+
+/// Iterator through proximity pairs.
+pub struct ProximityPairs<'a, P: Point + 'a, M: 'a, T: 'a> {
+    objects: &'a UidRemap<CollisionObject<P, M, T>>,
+    pairs:   Iter<'a, Entry<Pair, Box<ProximityDetector<P, M> + 'static>>>
+}
+
+impl<'a, P: 'a + Point, M: 'a, T: 'a> ProximityPairs<'a, P, M, T> {
+    #[doc(hidden)]
+    #[inline]
+    pub fn new(objects: &'a UidRemap<CollisionObject<P, M, T>>,
+               pairs:   Iter<'a, Entry<Pair, Box<ProximityDetector<P, M> + 'static>>>)
+               -> ProximityPairs<'a, P, M, T> {
+        ProximityPairs {
+            objects: objects,
+            pairs:   pairs
+        }
+    }
+}
+
+impl<'a, P: Point, M, T> Iterator for ProximityPairs<'a, P, M, T> {
+    type Item = (&'a CollisionObject<P, M, T>,
+                 &'a CollisionObject<P, M, T>,
+                 &'a ProximityAlgorithm<P, M>);
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        match self.pairs.next() {
+            Some(p) => {
+                let co1 = &self.objects[p.key.first];
+                let co2 = &self.objects[p.key.second];
+
+                Some((&co1, &co2, &p.value))
+            }
+            None => None
+        }
+    }
+}